LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-from-initial-about-blank-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL
CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL


Harness Error (TIMEOUT), message = null

TIMEOUT navigate() from <iframe> still on initial about:blank works Test timed out
